requirements.txt
streamlit
python-docx
google-generativeai
langchain-google-genai
langchain
faiss-cpu
pypdf
